requirements.txt
# Core framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Multi-agent framework
pyautogen>=0.2.0

# API clients
openai>=1.0.0
anthropic>=0.25.0

# Web search and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=2.1.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Data analysis
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.0

# Web interface
gradio>=4.0.0
